Remove structured_name from spec
This field was created to fulfill a function that it is not doing in it's current state. The thought was to have a field that was automatically generated based on the templated info already present in the BCO. I believe at this stage that it is out side the scope of what we need to develop working examples. I think we should re-examine this at a later date. I am going to remove this field for the time being.

Apply #14